Added
Added the possibilty to set custom radar images for ships and stations.
For ships, the standard arrow is displayed if the ship is unidentified,
the custom image is displayed (shaded by the faction color) if
identified. For stations, the image is displayed.

This allows for quick identification of ships by all officers in the
future as no custom icon is designed for any ship as for now.