The Edge Of Space - suggested changes
After finding the issue with setBeam instead of setBeamWeapon I thought
I'd fix a couple of spelling issues and try to address a couple of
points in the mission where I've seen different groups of players
getting confused.

1) After docking with EOS Scope it seems to take them a while to realize
they have done what they need to here and that they should return to
Central Command. The changes to the script and addition of a "Technical
Officer" leading an away team should help push them along.

2) I've seen two groups sit scratching their heads as to where sector C5
would be. They both figured it out but I maybe allowing a prod in that
direction by giving them a heading could be nice.