More elegant timer for inspection.
Now using an actual timer rather than moving a ship around to trigger
messages related to the inspection of EOS Scope.

I couldn't resist adding a little more flavour text for the technician,
especially to cater for a scenario where the players are no longer
docked when the inspection of EOS Scope is complete.