Fix timing scenario
Basically our team fought killed the same ship 10 times, got super boring to the point where we gave up. (The rescue ship took way to long) So I changed the spawn of the rescue ship to closer to save others from the headache we had. (Super fun up this point by the way)

1-4 attacks would be perfect.