Adding OptionsMenu for Hotkeys
this menu is situated in src/menus/HotKeyMenu.cpp and shows HotKeys
depending on the chosen Category. In-game it is accessible via the
OptionsMenu. It also allows to change values, this is why additional
functions had to be added to hotkeyConfig for set Hotkey values and
for readig all available fields.
I still need to implement a test for validitiy of the chosen Hotkey.