Rewrite Science sidebar pager
The sidebar pager for detailed scan info previously appeared only
when the target object was fully scanned, or if it had a
description. Always show it, leaving the panels without scan data
reporting No Data, so we can always show the signal data.

The Signal page of the sidebar exists and populates only if the
ship has_signal_radar.

"Noise" from distance on unscanned objects >5U away affects the
data's precision. Values >1.0 add noise to the visualization.
Scanned or nearby objects do not distort their data.

This moves the signal detail values off of the radar itself.