Add short range radar range preference
Add `server_config_short_range_radar_range` to PreferencesManager
and modify short-range radar view and screens accordingly.

Values are bounded to a range of 500 and 20000 (0.5U to 20U).

The default value remains 5000.

Nebula fog of war behavior now relies on this setting.

- If set to greater than 5000, helm, weapons, tactical, and
  single pilot use NebulaFogOfWar as science does.
- Friendly shared radar ranges on relay scale with the value.
- If close enough to a nebula, science can see through it up to
  the value.

This value also affects probe radar range.

This does _not_ add the setting to the server creation screen.
It is set only in options.ini (for now).