Mostly functional `view_rotation` for `RadarView`.
`RadarView` `view_position` now has a twin `view_rotation`.

`SpaceObject`'s render to radar functions now take a rotation argument as well.

Most of the draw functions in `RadarView` now  use the preexisting but updated `worldToScreen` function rather than calculating the position individually.  It may be appropriate to have `radar_screen_center` and `scale` be member variables since they are used all over the place so they do not need to be recalculated all the time.

`RenderView::DrawTargetProjections` function is the primary function which still needs updated.   This function draws the expected path of homing missiles.  I believe this function also draws movement predictions for targeted ships.

Another primary problem to solve is manual targeting.  Should the manual targeting rotate with the ship or with the view?  The manual targeting indicator also still needs updated to respect `view_rotation`.

The last major thing which needs updated is the science rgb wave around the edge of the radar.

After that some testing for whether something was missed.  And the option/control for whether to view according to "space north" or ship direction (Should it be a button or a config setting?).