Convert Single Pilot to glass cockpit
On 3D-capable devices, allow and default to replacing the static
background with a 3D viewport, of either a forward chase camera view
or a first-person view (using the MAIN_SCREEN FIRST_PERSON keybind
and respecting the first_person preferences file option). The radar
is semi-transparent and can be resized.

Allow toggling the 3D viewport to a normal opaque background.

Add hotkey bindings to toggle radar size (SINGLE_PILOT category,
TOGGLE_RADAR_SIZE, G default) and visibility (TOGGLE_VIEWPORT,
V default).

On devices not capble of rendering the 3D viewport, default to the
normal view with a static background.