Add getOnly PreferencesManager function
PreferencesManager::get() sets a value for the key, even if no value is passed.
This results in the getter setting a value, which isn't always desirable.

In this case, the key is written to options.ini. This is usually fine since it
self-documents what settings are available. But this makes deprecating settings
difficult, because there's no way to check whether a deprecated value is set
without also adding it to options.ini.