Fix client/server desync for tubes
Fixes #1048
if startLoad or startUnload is called within 0.5 seconds of the last
update of delay the new value wasn't sent to clients, resulting in
different paths later in update function, resulting in turn with
different tube states.

There probably is a risk of a similar out of sync with changing system
effectiveness, but the window for that is likely extremely minimal, the
consequences low, and as yet has not been observed in testing.