Remove extraneous get/setWeaponTubeSize functions
PR #933 added the current get/setTubeSize to spaceShip.

PR #679 added get/setWeaponTubeSize to spaceShip, as well as
setTubeSize to shipTemplate.

setTubeSize and setWeaponTubeSize are redundant. setTubeSize is
exposed to scripting, setWeaponTubeSize is not.

get/setWeaponTubeSize is not used anywhere in EmptyEpsilon. Because
it is redundant and unused, it should be removed.

Closes #1025.