Add direct ship rotation controls on Helm bindings
Move TURN_LEFT and TURN_RIGHT handling to onDraw and use
isKeyPressed to rotate the ship using `commandTurnSpeed()`, and
only when the keys are held down. This replaces the current Helms
behavior of relying on repeating keystroke behavior and sending
repeating `commandTargetRotation()` calls.

This needs testing to confirm whether it conflicts with joystick
ROTATE input, which also relies on `commandTurnSpeed()`.

Add optional Cockpit View keybinds to provide keyboard input more
similar to mouse-click input on the wheel, where the ship follows
the target.