Add hotkey remapping options (#1050)
* Adding OptionsMenu for Hotkeys

this menu is situated in src/menus/HotKeyMenu.cpp and shows HotKeys
depending on the chosen Category. In-game it is accessible via the
OptionsMenu. It also allows to change values, this is why additional
functions had to be added to hotkeyConfig for set Hotkey values and
for readig all available fields.
I still need to implement a test for validitiy of the chosen Hotkey.

* Improving implementation, Test of Keys

This commit adds a Test for user set HotKeys to prevent setting of keys
that are not in the smfl-list. Also I changed the key string comparison
to case insensitive comparison.

* Removed boost dependency

boost function iequals for insensitive string comparison has been
replaced with function iequals in hotkeyConfig.

* Easier string case isensitive comparison

removes function iequals from hotkeyConfig

* Make Basic untachable for user

So far Basic Hotkeys are hardcoded and hotkeyConnfig values have no
influence. Therefore, user should not be able to set those.

* Add hotKeyMenu to CMakeLists

* Integrate hot key config to paged options

* Paginate long lists of hotkeys, remove redundant elements

* Paging buttons deactive when not needed

* Consistent case for 'hotkey'

* Convert a system headers include to a local include

* Style and consistency edits

* Abstract hardcoded values, rename frame

At least use consts for hardcoded values to make them easier to
modify, and rename `frame` to `rebinding_ui` and `updateHotkeys` to
`saveHotkeys`.

* Convert pager buttons to arrows

* Attempt to implement onKey binding

* Tweak key field sizes for long key names

* Base rebinding UI frame height on row count const

* Make rows/cols of rebinds an autolayout

* Document GuiHotkeyBinder

Co-authored-by: greatestKeks <wolf.flo11235@gmail.com>
Co-authored-by: oznogon <garrett+ozgh@github.com>