Add remappable voice chat keys (#1288)
Add HOTKEY.BASIC.VOICE_CHAT_ALL and HOTKEY.BASIC.VOICE_CHAT_SHIP
for broadcasting voice to the server and player's ship, respectively.
Set default hotkey bindings to their current hardcoded keys of
Backspace and Tilde, respectively.

The hotkey menu in Options doesn't have access to the
NetworkAudioRecorder object in `main()` where the keys are bound,
so they're put in the BASIC hotkey category, which can't be rebound
via the hotkeys menu. However, the keys can be rebound in
`options.ini`.