Roam near position if no target position is set
If no order_target_position is set, a ship's orders are set to AI_Roaming, and
no targets are within 50U, ships following roaming orders will move towards 0,0
because that's the default value if no target position is unassigned.

When in this state, instead get and use a random target position within 30U of
the roaming ship's current position.

This significantly changes the behavior of ships more than 50U away from an
enemy. Scenarios that rely on this behavior might be affected.

Fixes #355.