Addition of translation tags for .ogg sound files and creation of the "scenariosSounds" folder
After discussing it with Xansta and testing what BlueShadow suggested, I added the translation tags for the sound files of the scenario. In order to avoid cluttering the "resources" folder I have created a "scenariosSounds" folder in which will be all the sound files specific to the scenarios. I will also modify the links of the other scenarios with sound, already existing (48, 54, 55 and 62) in order to bring them all together in the "ScenariosSounds" folder.