Stencil-based radar (#1413)
* Radar with stencil buffer

Drawing the radar screen with a stencil buffer to do the masking technique.

This drops the required RT to... 1, and avoids a slew of blend state dance.

* Ensures the RT is activated when interleaving gl calls.

* Make sure we exit with the RT deactivated.

* Process space objects in two passes

This avoids large memory operations.

* Clean ups

Cleaning up comments mostly.

* Allow radar to be drawn straight onto the backbuffer

SFML really wants a packed stencil/depth buffer to be available to use the stencil, however that is not the case on all devices, esp. in the Android realm.

When we fail to acquire a render texture, fallback on the backbuffer.

* Never use a render texture on Android devices.

* Forego RTs on ES systems

This comes from how SFML set up GL ES devices internally.

* Check Android System, not 'opengl es'

It appears `SFML_OPENGL_ES` is an internal define, rip.