Fix bad glFrustum on resize
When the 3D screen is displayed as a side screen, and the user resizes in a way that will hide it, there's one frame where the viewport will try to render in a 0-width viewport.

This leads to a bad gl state (glFrustum namely complains), so we just skip the draw when hitting that case.