Touch friendly main screen
Updated controls so that tapping the radar will bring up the tactical/long range view, and tapping the screen on the left rotates left and on the right rotates right.
When the radar is full screen, tapping onto it will switch between either tactical/long range, and tapping 'outside' will return back to the front view.