[CpuShip] Prevent orderAttack() from targeting non-hostiles
Because CpuShip:orderAttack() doesn't check if the target is
hostile, it can be used to attempt to attack non-hostile targets.
SpaceShips won't fire at non-hostile targets, so the result is
poorly defined.

When using orderAttack(), check first if the target is hostile.
If it isn't, write a warning to the game log and ignore the order.