modify identical entries on the other scenarios of Xansta
I apologize if I do too much PR for the translation of a single scenario.
As there are a lot of entries and they are not easy to translate (size, direction etc) and moreover I have to modify identical entries on the other scenarios of Xansta (size too big) I prefer to push several PR rather than making multiple commits. If it's annoying let me know @daid  Thanks