Update scenario_48_visitors.lua
With @xansta, we just realized yesterday that all the updates that Xansta has carried out on these scenarios over the last 6 months have been done without taking into account the updates on your Daid repo. This includes the major work of standardizing and adding 4000 tags and translation contexts.
In order not to lose all this work, we will compare the 2 file versions and update so that there is only 1 identical version between your daid repo and the xansta file.
Here is this first scenario in which I simply put back the translation contexts already added in the past in order to maintain standardization and make the translators' task easier.