tutorials: start specific tutorial before autoconnect
Changes the behaviour of the "tutorial" preference in options.ini or as command line argument. You can now specify which tutorial is started by the name of the tutorial file (e.g. "tutorial=tutorial/00_all.lua" for all tutorials).

If the autoconnect option is defined at the same time, the tutorial is started first and the connection to the server happens after the tutorial is finished.

This allows for example to show the helms tutorial before connecting to the helms station of a running server.