Add F1 keyboard references to cinematic, top-down views (#2440)
* Add F1 keyboard references to cinematic, top-down views

* Resolve forward-declaration compile errors

Adding `helpOverlay` to `cinematicViewScreen` and `topDownScreen`
causes compilation to fail with `invalid use of incomplete type ‘class GuiPanel’`
errors when setting visibility on the `helpOverlay`'s `frame`.

Per daid in #2402 and #2440, to reduce header fanout, resolve this
by including `gui/gui2_panel.h` in the screens' source files instead
of in `helpOverlay.h`.