Prevent uninitialized position and size of new GM chat dialogs
When a comms entity is being hailed by or has an open channel with
a GM that isn't created via the GM screen "Hail ship" button, it
can be created via GameMasterScreen::getChatDialog with an
uninitialized position and size. This can result in an unusuable
dialog and infinite GUI layout loop.

Assign a default position (0, 0) and size (300, 300) to any dialog
created via getChatDialog.