Handle backtracking from hotkey menu to options menu origin
Pass the options menu origin to the hotkey menu, so that when
exiting the options menu, you return to the correct screen.

Otherwise, it's possible to escape to the main menu while the
client/server is in a game by entering the options menu from ship
selection, then entering the hotkey menu, then exiting the hotkey
menu and options menu.