Split Hull and entity Health into components
Separate the presence of a Hull (a marker indicating ship-like
integrity) from an entity's health by creating a new Health
component. This allows the creation of entities that have health
and can be weapons targeted, but otherwise aren't treated as
having repairable ship hulls for the purposes of AI targeting or
threat estimates, and aren't treated as ships by Science or Relay,

This remains backward compatible in scripting by translating
set/getHull... Lua functions into respective set/getHealth...
functions, and adding the Hull component if the set/getHull...
functions are used.

In the process, this makes other related significant changes:

- A new FactionRelationship::None (`none` in Lua) indicates when
  a target lacks a Faction component. These previously reported as
  Neutral, which could affect targeting.
- Beams can now fire at any non-Friendly FactionRelationship. This
  previously inconsistently excluded Neutral targets from weapons,
  which could target Neutrals with missiles but couldn't fire beams
  at them.